View reservation details route complete
